Extraced "tasks.releaseDependencies()" function to a file. [skip ci]
